+++ public/modules/custom/grants_profile/src/Form/GrantsProfileFormPrivatePerson.php
@@ -280,38 +280,6 @@ you can do that by going to the Helsinki-profile from this link.', [], $this->tO
     );
   }
 
-  public function handleViolations(
-    ComplexDataDefinitionBase $grantsProfileDefinition,
-    array $grantsProfileContent,
-    FormStateInterface &$formState,
-    array $form,
-    array $addressArrayKeys,
-    array $officialArrayKeys,
-    array $bankAccountArrayKeys
-  ) {
-    // Create data object.
-    $grantsProfileData = $this->typedDataManager->create($grantsProfileDefinition);
-    $grantsProfileData->setValue($grantsProfileContent);
-    // Validate inserted data.
-    $violations = $grantsProfileData->validate();
-    // If there's violations in data.
-    if ($violations->count() == 0) {
-      // Move addressData object to form_state storage.
-      $freshStorageState = $formState->getStorage();
-      $freshStorageState['grantsProfileData'] = $grantsProfileData;
-      $formState->setStorage($freshStorageState);
-      return;
-    }
-    $this->reportValidatedErrors(
-      $violations,
-      $form,
-      $formState,
-      $addressArrayKeys,
-      $officialArrayKeys,
-      $bankAccountArrayKeys
-    );
-  }
-
   /**
    * {@inheritdoc}
    */
