+++ public/modules/custom/grants_profile/src/Form/GrantsProfileFormUnregisteredCommunity.php
@@ -226,22 +226,11 @@ you can do that by going to the Helsinki-profile from this link.', [], $this->tO
     $grantsProfileDefinition = GrantsProfileUnregisteredCommunityDefinition::create(
       'grants_profile_unregistered_community');
     // Create data object.
-    $grantsProfileData = $this->typedDataManager->create($grantsProfileDefinition);
-    $grantsProfileData->setValue($grantsProfileContent);
-    // Validate inserted data.
-    $violations = $grantsProfileData->validate();
-    // If there's violations in data.
-    if ($violations->count() == 0) {
-      // Move addressData object to form_state storage.
-      $freshStorageState = $formState->getStorage();
-      $freshStorageState['grantsProfileData'] = $grantsProfileData;
-      $formState->setStorage($freshStorageState);
-      return;
-    }
-    parent::reportValidatedErrors(
-      $violations,
-      $form,
+    $this->handleViolations(
+      $grantsProfileDefinition,
+      $grantsProfileContent,
       $formState,
+      $form,
       $addressArrayKeys,
       $officialArrayKeys,
       $bankAccountArrayKeys
